backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery[redis]==5.4.0
redis==5.1.0
motor==3.6.0
pymongo==4.9.1
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
opencv-python-headless==4.10.0.84
mediapipe==0.10.18
librosa==0.10.2
google-generativeai==0.8.2
python-multipart==0.0.12
google-cloud-storage==2.18.2
numpy==1.26.4
